feat: Use `commits`, `lastRelease` and `nextRelease` passed by `semantic-release`
Use the same `commits`, `lastRelease` and `nextRelease` as other `semantic-release` plugin, instead of retrieving them locally with `conventional-commits-core`. That avoid discrepancy in the determination of next version, last release and commits to include in release note.

BREAKING CHANGE: Expect to be passed `commits`, `lastRelease` and `nextRelease` in `options`. Require `semantic-remantic` `>=9.0.0`.
BREAKING CHANGE: Do not wrap unexpected errors in a `@semantic-release/error`. This way, in case of unexpected error (missing preset for example) `semantic-release` will fail and return with an error code