chore: use empathic for package traversal
Since we disable normalisation, we're essentially traversing upwards for
the `package.json` and reading it as-is. We can use a much faster and
smaller library for this (empathic in this case).

Since empathic is generic (i.e. can find any file upwards), it'll also
de-dupe nicely in dependency trees as adoption increases.